Started ui

+++ src/main/model/Transaction.java
@@ -6,14 +6,14 @@ public class Transaction {
     //fields
     private String title;
     private LocalDate date;
-    private int amount;
+    private double amount;
     private Category transactionType;
     private String description;
 
     //constructor
     // REQUIRES: amount > 0
     // EFFECTS: creates earning  with a generic title, date, amount as 0, and category type.
-    public Transaction(String name, LocalDate date, int amount, String description, Category type) {
+    public Transaction(String name, LocalDate date, double amount, String description, Category type) {
         this.title = name;
         this.date = date;
         this.amount = amount;
@@ -21,7 +21,7 @@ public class Transaction {
         this.description = description;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
@@ -37,7 +37,7 @@ public class Transaction {
         return date;
     }
 
-    public void setAmount(int x) {
+    public void setAmount(double x) {
         this.amount = x;
     }
